Remove the useless Openshift::ContainerManagerMixin
Kubernetes has a ContainerManagerMixin which is shared by the
Kubernetes::ContainerManager and Openshift.  Nothing else includes the
Openshift::ContainerManagerMixin other than the
Openshift::ContainerManager so it is a useless mixin
